requirements.txt
pandas
pyyaml
sqlalchemy
psycopg2-binary
streamlit
plotly
